+++ sdk/keyvault/Azure.Security.KeyVault.Keys/samples/Sample3_GetKeys.cs
@@ -52,11 +52,11 @@ namespace Azure.Security.KeyVault.Keys.Samples
             // Let's list the keys and print their types.
             // List operations don't return the keys with key material information.
             // So, for each returned key we call GetKey to get the key with its key material information.
-            IEnumerable<Response<KeyBase>> keys = client.GetKeys();
-            foreach (KeyBase key in keys)
+            IEnumerable<Response<KeyProperties>> keys = client.GetKeys();
+            foreach (KeyProperties key in keys)
             {
                 Key keyWithType = client.GetKey(key.Name);
-                Debug.WriteLine($"Key is returned with name {keyWithType.Name} and type {keyWithType.KeyMaterial.KeyType}");
+                Debug.WriteLine($"Key is returned with name {keyWithType.Properties.Name} and type {keyWithType.KeyMaterial.KeyType}");
             }
 
             // We need the Cloud RSA key with bigger key size, so you want to update the key in Key Vault to ensure
@@ -72,8 +72,8 @@ namespace Azure.Security.KeyVault.Keys.Samples
 
             // You need to check all the different versions Cloud RSA key had previously.
             // Lets print all the versions of this key.
-            IEnumerable<Response<KeyBase>> keysVersions = client.GetKeyVersions(rsaKeyName);
-            foreach (KeyBase key in keysVersions)
+            IEnumerable<Response<KeyProperties>> keysVersions = client.GetKeyVersions(rsaKeyName);
+            foreach (KeyProperties key in keysVersions)
             {
                 Debug.WriteLine($"Key's version {key.Version} with name {key.Name}");
             }