+++ sdk/keyvault/Azure.Security.KeyVault.Keys/src/Key.cs
@@ -6,35 +6,64 @@ using System.Text.Json;
 namespace Azure.Security.KeyVault.Keys
 {
     /// <summary>
-    /// Key is the resource consisting of name, <see cref="JsonWebKey"/> properties and its attributes inherited from <see cref="KeyBase"/>.
+    /// <see cref="Key"/> is the resource consisting of a value and its <see cref="Properties"/>.
     /// </summary>
-    public class Key : KeyBase
+    public class Key : IJsonDeserializable, IJsonSerializable
     {
+        private const string KeyPropertyName = "key";
+
+        internal Key()
+        {
+            Properties = new KeyProperties();
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="Key"/> class.
+        /// </summary>
+        /// <param name="name">The name of the key.</param>
+        public Key(string name)
+        {
+            Properties = new KeyProperties(name);
+        }
+
         /// <summary>
         /// As of http://tools.ietf.org/html/draft-ietf-jose-json-web-key-18
         /// </summary>
         public JsonWebKey KeyMaterial { get; set; }
 
-        internal Key() { }
-
         /// <summary>
-        /// Initializes a new instance of the Key class.
+        /// Gets or sets the attributes of the <see cref="Key"/>.
         /// </summary>
-        /// <param name="name">The name of the key.</param>
-        public Key(string name) : base(name) { }
+        public KeyProperties Properties { get; }
 
-        internal override void ReadProperties(JsonElement json)
+        internal virtual void ReadProperty(JsonProperty prop)
         {
-            KeyMaterial = null;
-
-            if (json.TryGetProperty("key", out JsonElement key))
+            switch (prop.Name)
             {
-                KeyMaterial = new JsonWebKey();
-                KeyMaterial.ReadProperties(key);
-                ParseId(KeyMaterial.KeyId);
-            }
+                case KeyPropertyName:
+                    KeyMaterial = new JsonWebKey();
+                    KeyMaterial.ReadProperties(prop.Value);
+                    Properties.ParseId(KeyMaterial.KeyId);
+                    break;
 
-            base.ReadProperties(json);
+                default:
+                    Properties.ReadProperty(prop);
+                    break;
+            }
         }
+
+        internal virtual void WriteProperties(Utf8JsonWriter json)
+        {
+        }
+
+        void IJsonDeserializable.ReadProperties(JsonElement json)
+        {
+            foreach (JsonProperty prop in json.EnumerateObject())
+            {
+                ReadProperty(prop);
+            }
+        }
+
+        void IJsonSerializable.WriteProperties(Utf8JsonWriter json) => WriteProperties(json);
     }
 }