+++ sdk/keyvault/Azure.Security.KeyVault.Secrets/ChangeLog.md
@@ -6,6 +6,7 @@
 
 - `SecretBase` has been renamed to `SecretProperties`.
 - `Secret` and `DeletedSecret` no longer extend `SecretProperties`, but instead contain a `SecretProperties` property named `Properties`.
+- `SecretClient.Update` has been renamed to `SecretClient.UpdateProperties`.
 - `SecretProperties.Vault` has been renamed to `SecretProperties.VaultUri`.
 
 ## 4.0.0-preview.1 (2019-06-28)